requirements.txt
psycopg2
django
djangorestframework
requests
gunicorn
dj-database-url
whitenoise
googlemaps
elasticsearch
django-haystack
